Se crea usuarios conservando todos los datos correctamente

+++ League planner/Jugador.cs
@@ -13,13 +13,13 @@ namespace League_planner
         public string ApellidoPaterno { get; set; }
         public string ApellidoMaterno { get; set; }
         public DateTime FechaDeNacimiento { get; set; }
-        public int Telefono { get; set; }
+        public long Telefono { get; set; }
         public string Email { get; set; }
         public int Equipo { get; set; }
         public bool Eliminado { get; set; }
         public string NombreCompleto;// { get => Nombre + " " + ApellidoPaterno + " " + ApellidoMaterno; } 
 
-        public Jugador(long id, string nombre,string apellidoPaterno,string apellidoMaterno,DateTime nacimiento,int tel,string mail, bool eliminado, int equipo)
+        public Jugador(long id, string nombre,string apellidoPaterno,string apellidoMaterno,DateTime nacimiento,long tel,string mail, bool eliminado, int equipo)
         {
             this.id = id;
             this.Nombre = nombre;
@@ -32,7 +32,7 @@ namespace League_planner
             Eliminado = eliminado;
         }
 
-        public Jugador(string nombre, string apellidoPaterno, string apellidoMaterno, DateTime nacimiento, int tel, string mail, bool eliminado, int equipo)
+        public Jugador(string nombre, string apellidoPaterno, string apellidoMaterno, DateTime nacimiento, long tel, string mail, bool eliminado, int equipo)
         {
             this.Nombre = nombre;
             this.ApellidoPaterno = apellidoPaterno;
